+++ coffeenet-autoconfigure/src/test/java/coffee/synyx/autoconfigure/navigation/CoffeeNetWebPropertiesTest.java
@@ -1,0 +1,23 @@
+package coffee.synyx.autoconfigure.navigation;
+
+import org.junit.Test;
+
+import static org.hamcrest.Matchers.is;
+
+import static org.junit.Assert.assertThat;
+
+
+/**
+ * @author  Tobias Schneider - schneider@synyx.de
+ */
+public class CoffeeNetWebPropertiesTest {
+
+    @Test
+    public void testDefaultValues() {
+
+        CoffeeNetWebProperties sut = new CoffeeNetWebProperties();
+
+        assertThat(sut.getLogoutPath(), is("/logout"));
+        assertThat(sut.getProfileServiceName(), is("profile"));
+    }
+}